Unit test for Circle
There are some bugs:

The segment is outside the circle, it will also return true for isIntersect.

The shape that does not intersect can pass the test, but if group these no intersect shape to a group, there is some bug. For all shapes in the group are not intersect with the circle, it will also return true.

I have tried to fix it, but I can't find where the bug is.